feat: Add About section components including Meet Team, Our Story, Student Adda Drives, Success Story, and Why Choose Us
